Exclude exception from notifier content
The exception trace can get very long, which will not be useful in some
of the notifier channels (E.G receiving an SMS with a big wall of text is
not very helpful). Also some channels might limit the amount of text that
can be sent which will result in api errors.