fix: don't auto-release lock
The existing behaviour auto-released the lock once
the `schedule:run`command finished execution.

For short running tasks (~1 second or less), auto-releasing the lock can
lead to a case where if a cron starts slightly later on another server,
the lock from the first invocation has already been deleted,
and the task is executed for a second time.

This change relies on the DEFAULT_TTL constant to tidy-up locks rather than
doing so in this extension.